// Kulthia Dasboard/src/app/core/models/raw.model.ts
export interface PurchaseRawMaterialModel {
    usersData: any;
    usersDetails: any;
  
    PurNo: any;
    CashPurchase:any;
    PurDate: any;
    BillDate: any;
    VendBill:String;
    PartyID:String;
    Address: any;
    DueDate: any;
    PurAc: String;
    Narr: String;
    GrossAmt: any;
    cgst: any;
    sgst: any;
    RndOffAmt: any;
    TotAmt: any;
 
  
}